Thread deletion button now works.

+++ database/migrations/2019_09_04_193420_add_category_id_to_threads.php
@@ -14,7 +14,7 @@ class AddCategoryIdToThreads extends Migration
     public function up()
     {
         Schema::table('threads', function (Blueprint $table) {
-            $table->decimal('categoryId');
+            $table->integer('categoryId');
         });
     }
