+++ database/migrations/2019_12_10_070312_change_category_name_to_category_id_on_threads.php
@@ -15,7 +15,7 @@ class ChangeCategoryNameToCategoryIdOnThreads extends Migration
     {
         Schema::table('threads', function (Blueprint $table) {
             $table->dropColumn('categoryName');
-            $table->decimal('categoryId');
+            $table->integer('categoryId');
         });
     }
